Applied patch by Steve Grubb:
kfile-plugins/jpeg/exif.cpp
process_EXIF is called without verifying that itemlen >= 8. If you look at the
called function, it accesses the 8th element in the data array. The same goes for
the process_SOFn function. For both of these, I placed a FIXME tag since I don't
know how you really want to handle it. The code does check for itemlen > 2, but
that's not sufficient for these 2 cases.

kfile-plugins/jpeg/kfile_setcomment.cpp has a couple of memory leaks.

svn path=/trunk/kdegraphics/kfile-plugins/; revision=376401