Update the PDF plugin to use Poppler, instead of relying on xpdf's pdfinfo.
As discussed on kde-devel, this relies on a new library that hasn't
actually been released yet. However it seems that it is being actively
developed, is a lot more sane than trying to exec a binary and
parse the output, and offers a lot more expansion capability.

There is also a philosphical issue - if Poppler is worth supporting in the
longer term, we should get behind it early.

svn path=/trunk/kdegraphics/kfile-plugins/; revision=401567