Check in an initial version of the kfile (meta-data) plugin for the XML Paper Specification format, which is an emerging Microsoft format for e-documents. The format is broadly similar in capabilities to what most people use PDF for. The approach in this document would probably work for similarly packaged files, including some of the stuff from Microsoft's Office 2007 format.
This plugin isn't the greatest example of XML handling,
but it does work on almost all the files I tried it on.

It is known to crash on some files - those that use
the interleaved Zip format. I just don't know how to
protect from this case. It also won't work on anything
that uses digital signatures (just not handled - we don't
have enough XAML tools) and also won't work on zip 
files that uses the Zip64 extensions (because KZip can't
handle that).

svn path=/trunk/KDE/kdegraphics/kfile-plugins/; revision=582427